backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# RBAC
casbin==1.42.0
casbin-sqlalchemy-adapter==0.2.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# A2A Protocol
httpx==0.26.0

# WebSocket
websockets==12.0

# Utils
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Type hints
annotated-types==0.6.0
